requirements.txt
certifi==2021.5.30
chardet==4.0.0
dnspython==1.16.0
idna==2.10
loguru==0.5.3
numpy==1.20.3
pandas==1.2.4
pymongo==3.11.4
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
requests-random-user-agent==2021.5.29
Shapely==1.7.1
six==1.16.0
urllib3==1.26.5
